Added challenge & updated responsiveness
Added Compound Interest Calculator that I made a few months ago as a standalone personal project.

Added a medium sized card popup... considering making a single popup for all cards rather than expanding the cards themselves... this will make it easier to style and animate the rest of the website to go along with the popup.